refactor: refactor configs, data-fetchers, services, utils
