+++ app/scripts/HeatmapTiledPixiTrack.js
@@ -23,9 +23,7 @@ import AxisPixi from './AxisPixi';
 // Services
 import { tileProxy } from './services';
 
-import { GLOBALS, HEATED_OBJECT_MAP } from './configs';
-
-import { NUM_PRECOMP_SUBSETS_PER_2D_TTILE } from './configs/dense-data-extrema-config';
+import { GLOBALS, HEATED_OBJECT_MAP, NUM_PRECOMP_SUBSETS_PER_2D_TTILE } from './configs';
 
 const COLORBAR_MAX_HEIGHT = 200;
 const COLORBAR_WIDTH = 10;
