+++ app/scripts/configs/available-for-plugins.js
@@ -81,8 +81,7 @@ import ViewportTrackerVertical from '../ViewportTrackerVertical';
 
 // Factories
 import ContextMenuItem from '../ContextMenuItem';
-import DataFetcher from '../DataFetcher';
-import LruCache from '../factories/LruCache';
+import LruCache from './LruCache';
 
 // Services
 import * as services from '../services';
@@ -92,9 +91,8 @@ import ChromosomeInfo from '../ChromosomeInfo';
 import SearchField from '../SearchField';
 
 // Data Fetchers
-import GBKDataFetcher from '../data-fetchers/genbank-fetcher';
-import LocalDataFetcher from '../data-fetchers/local-tile-fetcher';
-import getDataFetcher from '../data-fetchers/get-data-fetcher';
+import { DataFetcher, GBKDataFetcher, LocalDataFetcher } from '../data-fetchers';
+import getDataFetcher from './get-data-fetcher';
 
 import { version } from '../../../package.json';
