+++ app/scripts/configs/get-data-fetcher.js
@@ -1,7 +1,5 @@
-import GBKDataFetcher from './genbank-fetcher';
-import LocalDataFetcher from './local-tile-fetcher';
-import DataFetcher from '../DataFetcher';
-import { AVAILABLE_FOR_PLUGINS } from '../configs';
+import { DataFetcher, GBKDataFetcher, LocalDataFetcher }  from '../data-fetchers';
+import { AVAILABLE_FOR_PLUGINS } from '.';
 
 const getDataFetcher = (
   dataConfig,
