+++ app/scripts/data-fetchers/index.js
@@ -1,0 +1,3 @@
+export { default as DataFetcher } from '../data-fetchers/DataFetcher';
+export { default as GBKDataFetcher } from '../data-fetchers/genbank-fetcher';
+export { default as LocalDataFetcher } from '../data-fetchers/local-tile-fetcher';
